requirements.txt
flask 
flask-cors 
diffusers 
torch 
torchvision 
transformers 
numpy 
pillow 
opencv-python 
flask-ngrok